Issue-1001: Move CronTaskController to strongbox-web-core and get rid of strongbox-cron-rest.
