+++ strongbox-web-core/pom.xml
@@ -294,7 +294,15 @@
             <type>test-jar</type>
             <scope>test</scope>
         </dependency>
-        
+
+        <dependency>
+            <groupId>${project.groupId}</groupId>
+            <artifactId>strongbox-cron-api</artifactId>
+            <version>${project.version}</version>
+            <type>test-jar</type>
+            <scope>test</scope>
+        </dependency>
+
         <dependency>
             <groupId>${project.groupId}</groupId>
             <artifactId>strongbox-storage-api</artifactId>
