+++ strongbox-web-core/src/test/resources/etc/conf/strongbox-cron-tasks.xml
