+++ strongbox-web-core/src/test/resources/groovy/GroovyTask.groovy
